modify the tests for the new escrow
